backend/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
Faker==39.0.0
gunicorn==23.0.0
Markdown==3.10
packaging==25.0
psycopg2-binary==2.9.11
redis==7.1.0
sqlparse==0.5.5
tzdata==2025.3
celery==5.3.6
django-celery-results==2.5.1
drf-spectacular==0.27.1

